Add ENV_DIR support
Prepare buildstep to support the ENV_DIR feature that allows usinge the configuration variables during the compile phase. This change will need a change to dokku to put split the ENV file and insert it into the container as separate files.

This also makes the sourcing of the files in the .profile.d directory only take `*.sh` into consideration, matching the way that Heroku works.

See https://devcenter.heroku.com/articles/buildpack-api#bin-compile-usage for details on the ENV_DIR feature.

This implements #101 but also writes the config_vars to `00_config_vars.sh` so that defaults from the buildpacks get honored. That should be OK since the config_vars don't overwrite existing environment variables anymore. Removing this support is currently ill advised since some of the included buildpacks (Scala, Play, Grails) use this to set up paths and JVM options.

This commit is a combination of #99, #100 and #101 adjusted slightly for the new builder structure.